+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 1.0.1
 + Safe check for missing players when accepting invites
++ Actions now get synced when a player joins a team